Make a test file just for unicode; more to come
