+++ tests/test_resources.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
@@ -67,29 +66,6 @@ def test_path_part_params_are_available(harness):
         %(a)s
     """, '/foo/index.html.spt', '/foo;a=1;b;a=3/')
     assert response.body == "3\n"
-
-def test_utf8(harness):
-    expected = unichr(1758)
-    actual = harness.simple("""
-        "#empty first page"
-        [------------------]
-        text = unichr(1758)
-        [------------------]
-        %(text)s
-    """).body.strip()
-    assert actual == expected
-
-def test_utf8_char(harness):
-    expected = u'א'
-    actual = harness.simple(b"""
-        # encoding=utf8
-        "#empty first page"
-        [------------------]
-        text = u'א'
-        [------------------]
-        %(text)s
-    """).body.strip()
-    assert actual == expected
 
 def test_resources_dont_leak_whitespace(harness):
     """This aims to resolve https://github.com/whit537/aspen/issues/8.
